requirements.txt
absl-py==0.8.0
astor==0.8.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
firebase-admin==3.0.0
Flask==1.1.1
gast==0.2.2
gcloud==0.17.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-pasta==0.1.7
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.1
gunicorn==19.9.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jws==0.1.3
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.17.2
oauth2client==3.0.0
opt-einsum==3.1.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycryptodome==3.4.3
Pyrebase==3.0.27
python-bcrypt==0.3.2
python-jwt==2.0.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.9
stripe==2.37.2
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
